refac: move Reviews to components
