+++ src/components/ReviewTeacher.vue
@@ -82,9 +82,9 @@
   import Highcharts from "highcharts";
 
   import _ from 'lodash'
-  import Api from '../helpers/api'
-  import Utils from '../helpers/utils'
-  import MatriculaHelper from '../helpers/matricula'
+  import Api from '../scripts/helpers/api'
+  import Utils from '../scripts/helpers/utils'
+  import MatriculaHelper from '../scripts/helpers/matricula'
 
   Highcharts3D(Highcharts);
